* Removed the remaining Winsock code. * Changed the temp file logic for the worse.  Will implement fresh approach. * Various cosmetic work to remove traces of Windows.
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/contrib@738 67a373e5-eb02-0410-a15c-ee090a768436

+++ fcptools/fcpput/fcpput.c
@@ -234,7 +234,7 @@ static void parse_args(int argc, char *argv[])
         else if (!strcmp(argv[i], "-a"))
             maxAttempts = (++i < argc)
                         ? atoi(argv[i])
-                        : usage("missing maxAttempts value");
+                        : (int)usage("missing maxAttempts value");
         else if (!strcmp(argv[i], "-st"))
             maxSplitThreads = (++i < argc)
                         ? atoi(argv[i])
